C12

+++ C11/ex05/src/putnbr.c
@@ -6,7 +6,7 @@
 /*   By: avapaill <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/26 18:55:34 by avapaill          #+#    #+#             */
-/*   Updated: 2022/07/27 00:49:27 by avapaill         ###   ########.fr       */
+/*   Updated: 2022/07/27 13:06:31 by avapaill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	ft_putchar(char c)
 	write(1, &c, 1);
 }
 
-void	write_number(unsigned long nb)
+void	write_number(unsigned int nb)
 {	
 	if (nb < 10)
 	{
@@ -28,13 +28,13 @@ void	write_number(unsigned long nb)
 	write_number(nb % 10);
 }
 
-void	ft_putnbr(long nb)
+void	ft_putnbr(int nb)
 {
 	if (nb < 0)
 	{
 		write(1, "-", 1);
-		write_number((unsigned long) -1 * nb);
+		write_number((unsigned int) -1 * nb);
 	}
 	else
-		write_number((unsigned long) nb);
+		write_number(nb);
 }
